# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Shahebbazar API
  version: 1.1.0-sprint2
  summary: Rajshahi local business directory — Express API.
  description: |
    The Express API behind shahebbazar.com. The Next.js frontend calls it
    from the server when rendering pages; the provider dashboard, the
    analytics beacons and image uploads call it from the browser.

    **Browse it:** run the API and open `http://localhost:4000/api/docs`.
    **Postman:** File → Import → this file (`docs/openapi.yaml`).

    **Authentication is not built yet.** Every provider and admin
    endpoint is marked `TODO(auth)` in the code and is currently open.
    The customer (`/api/me/...`) endpoints act as the first active
    customer account until sign-in exists.
    They must not be deployed publicly until sessions and role checks
    exist. The provider endpoints take the business id in the path for
    now; with sign-in, it will come from the session.

    **Conventions**
    - All public responses contain approved records only.
    - `NUMERIC` and `COUNT(*)` values that PostgreSQL returns as big
      numbers arrive as **strings** (e.g. `"rating": "4.5"`); convert
      before doing arithmetic.
    - Stored image paths are relative (`uploads/…`) and are served by this
      API at `/uploads/…`.
    - Errors are `{ "error": "…" }`; form endpoints add `errors` keyed by
      field.
  contact:
    name: COS40005 Group 23
# No authentication exists yet, so no security scheme is declared. When
...
security: []
servers:
...
paths:
...
  /api/businesses/{slug}/payment-methods:
    get:
      operationId: getBusinessesBySlugPaymentMethods
      tags: [Public]
      summary: Payment methods a business accepts (display only)
      parameters:
        - $ref: "#/components/parameters/Slug"
      responses:
        "200":
          description: May be empty.
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/PaymentMethod" } }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/vendors/{vendorId}/payment-methods:
    get:
      operationId: getVendorsByVendorIdPaymentMethods
      tags: [Provider]
      summary: The business's payment methods, for editing
      parameters:
        - $ref: "#/components/parameters/VendorId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [vendor, methods]
                properties:
                  vendor: { $ref: "#/components/schemas/VendorSummary" }
                  methods: { type: array, items: { $ref: "#/components/schemas/PaymentMethod" } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      operationId: putVendorsByVendorIdPaymentMethods
      tags: [Provider]
      summary: Replace the whole set of payment methods
      description: Notes may not contain an 8+ digit run, so account or wallet numbers cannot be published.
      parameters:
        - $ref: "#/components/parameters/VendorId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [methods]
              properties:
                methods: { type: array, items: { $ref: "#/components/schemas/PaymentMethod" } }
      responses:
        "200":
          description: The saved set.
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/PaymentMethod" } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ------------------------------------------------------------ listings
...
  /api/admin/queue:
    get:
      operationId: getAdminQueue
      tags: [Admin]
      summary: Pending businesses and listings, and recent decisions
      responses:
        "200":
          description: Oldest pending first; last 20 decisions.
          content:
            application/json:
              schema:
                type: object
                required: [vendors, listings, recent]
                properties:
                  vendors: { type: array, items: { $ref: "#/components/schemas/PendingBusiness" } }
                  listings: { type: array, items: { $ref: "#/components/schemas/PendingListing" } }
                  recent: { type: array, items: { $ref: "#/components/schemas/AdminDecisionLog" } }
        "500": { $ref: "#/components/responses/ServerError" }
...
components:
  parameters:
    Slug:
      name: slug
      in: path
      required: true
      schema: { type: string }
      example: padma-view-restaurant
    VendorId:
      name: vendorId
      in: path
      required: true
      description: TODO(auth) — will come from the session.
      schema: { type: integer, minimum: 1 }
      example: 13
...
  responses:
...
    BadRequest:
      description: Invalid input.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: Not found (for public endpoints, also when not approved).
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    ServerError:
      description: Server error. Outside production, `detail` carries the database message.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        detail: { type: string }
        hint: { type: string }
...
    NumericString:
      type: string
      pattern: '^-?\d+(\.\d+)?$'
      description: A PostgreSQL NUMERIC or BIGINT, as a string.
    NullableString: { type: [string, "null"] }
    VendorStatus: { type: string, enum: [draft, pending, approved, rejected, suspended] }
...
    PaymentMethod:
      type: object
      required: [payment_method, payment_note]
      properties:
        payment_method: { type: string, enum: [cash, cash_on_delivery, bank_transfer, bkash, nagad, rocket, card] }
        payment_note: { type: [string, "null"], maxLength: 120 }
    VendorSummary:
      type: object
      required: [vendor_id, vendor_name, vendor_slug, vendor_status]
      properties:
        vendor_id: { type: integer }
        vendor_name: { type: string }
        vendor_slug: { type: string }
        vendor_status: { $ref: "#/components/schemas/VendorStatus" }
...
    PendingBusiness:
      type: object
      required: [vendor_id, vendor_slug, vendor_name, vendor_phone, vendor_updated_at, owner_name, owner_phone, owner_phone_verified, listing_count, photo_count, hours_days, owner_other_businesses, resubmitted]
      properties:
        vendor_id: { type: integer }
        vendor_slug: { type: string }
        vendor_name: { type: string }
        vendor_name_bn: { $ref: "#/components/schemas/NullableString" }
        vendor_description: { $ref: "#/components/schemas/NullableString" }
        vendor_phone: { type: string }
        vendor_address: { $ref: "#/components/schemas/NullableString" }
        vendor_cover_url: { $ref: "#/components/schemas/NullableString" }
        vendor_created_at: { type: string, format: date-time }
        vendor_updated_at: { type: string, format: date-time }
        area_name: { $ref: "#/components/schemas/NullableString" }
        owner_name: { type: string }
        owner_phone: { type: string }
        owner_phone_verified: { type: boolean }
        categories: { $ref: "#/components/schemas/NullableString" }
        listing_count: { type: integer }
        photo_count: { type: integer }
        hours_days: { type: integer, minimum: 0, maximum: 7 }
        owner_other_businesses: { type: integer }
        resubmitted: { type: boolean, description: "Rejected before and sent back by the owner." }
    PendingListing:
      type: object
      required: [listing_id, listing_title, listing_created_at, vendor_id, vendor_name, vendor_slug, vendor_status]
      properties:
        listing_id: { type: integer }
        listing_title: { type: string }
        listing_description: { $ref: "#/components/schemas/NullableString" }
        listing_price: { oneOf: [{ $ref: "#/components/schemas/NumericString" }, { type: "null" }] }
        listing_price_max: { oneOf: [{ $ref: "#/components/schemas/NumericString" }, { type: "null" }] }
        listing_price_unit: { $ref: "#/components/schemas/NullableString" }
        listing_min_order_qty: { type: [integer, "null"] }
        listing_created_at: { type: string, format: date-time }
        category_name: { $ref: "#/components/schemas/NullableString" }
        vendor_id: { type: integer }
        vendor_name: { type: string }
        vendor_slug: { type: string }
        vendor_status: { $ref: "#/components/schemas/VendorStatus" }
        photo_url: { $ref: "#/components/schemas/NullableString" }
    AdminDecisionLog:
      type: object
      required: [audit_id, audit_action, audit_target_type, audit_target_id, audit_created_at]
      properties:
        audit_id: { $ref: "#/components/schemas/NumericString" }
        audit_action: { type: string, enum: [vendor.approve, vendor.reject, listing.approve, listing.reject] }
        audit_target_type: { type: string, enum: [vendor, listing] }
        audit_target_id: { type: integer }
        reason: { $ref: "#/components/schemas/NullableString" }
        audit_created_at: { type: string, format: date-time }
        actor_name: { $ref: "#/components/schemas/NullableString" }
        target_label: { $ref: "#/components/schemas/NullableString" }